Fix parsing error introduced due to a change story html
Perhaps another element was introduced between the byline authored and
date elements. It messed up the parsing of the authored and date
elements when processing the byline div. This would produce a run-time
error.

It is fixed, for now, but I would prefer to apply some more robust
parsing. The markup is not all that suitible for that, however.
Of course, the best course of action would be to use an api, but hey,
not available to me right now :/